Combat Calculations Refactor

+++ src/TbsFramework/Nodes/Components/UnitMapAnimations.cs
@@ -21,9 +21,9 @@ public abstract partial class UnitMapAnimations : Node2D
 
     /// <summary>Set the unit's maximum health value to indicate on the map.</summary>
     /// <param name="value">New maximum health value.</param>
-    public abstract void SetHealthValue(int value);
+    public abstract void SetHealthValue(double value);
 
     /// <summary>Set the unit's current health value to indicate on the map.</summary>
     /// <param name="value">New current health value.</param>
-    public abstract void SetHealthMax(int value);
+    public abstract void SetHealthMax(double value);
 }
